+++ src/lib/AdminPage.svelte
@@ -198,7 +198,7 @@
 
 </script>
 
-<Modal title="Change Admin Username" bind:open={nameModal} autoclose>
+<Modal title="Change Admin Username" bind:open={nameModal} autoclose outsideclose>
     <p>
         <span>Please provide an updated name for {selectedAdmin.name} ({selectedAdmin.email})</span>
         <br>
@@ -208,21 +208,25 @@
     </p>
     
     <!-- TODO: CHANGE THESE COLOURS -->
-    <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminName(selectedAdmin)}>Confirm</Button>
-    <Button class="bg-red-200 hover:bg-red-300 text-black">Cancel</Button>
+    <svelte:fragment slot="footer">
+        <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminName(selectedAdmin)}>Confirm</Button>
+        <Button class="bg-red-200 hover:bg-red-300 text-black">Cancel</Button>
+    </svelte:fragment>
 </Modal>
 
-<Modal title="Notice for Revoking Admin Permissions" bind:open={adminFalseModal} autoclose>
+<Modal title="Notice for Revoking Admin Permissions" bind:open={adminFalseModal} autoclose outsideclose>
     <p>
         <span class="text-red-600">Are you sure you want to revoke the permissions of {selectedAdmin.name} ({selectedAdmin.email})?</span>
     </p>
     
     <!-- TODO: CHANGE THESE COLOURS -->
-    <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminPermissions(selectedAdmin)}>Yes</Button>
-    <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    <svelte:fragment slot="footer">
+        <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminPermissions(selectedAdmin)}>Yes</Button>
+        <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    </svelte:fragment>
 </Modal>
 
-<Modal title="Notice for Granting Admin Permissions" bind:open={adminModal} autoclose>
+<Modal title="Notice for Granting Admin Permissions" bind:open={adminModal} autoclose outsideclose>
     <p>
         <span>Giving a user admin status allows them to view and modify all employees in the system. Only do this if you know and trust this person.</span>
         <br>
@@ -231,11 +235,13 @@
     </p>
     
     <!-- TODO: CHANGE THESE COLOURS -->
-    <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminPermissions(selectedAdmin)}>Yes</Button>
-    <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    <svelte:fragment slot="footer">
+        <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => modifyAdminPermissions(selectedAdmin)}>Yes</Button>
+        <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    </svelte:fragment>
 </Modal>
 
-<Modal title="Notice for Deleting Admins" bind:open={deleteModal} autoclose>
+<Modal title="Notice for Deleting Admins" bind:open={deleteModal} autoclose outsideclose>
     <span>The following users will be removed from the database and will no longer have permissions to access the SLHC Employee Hearing Panel.</span>
     {#each selectedAdmins as admin}
         <p>{admin.name} ({admin.email})</p>
@@ -244,8 +250,10 @@
     <span class="text-red-600">Are you sure you want to remove these admins?</span>
     
     <!-- TODO: CHANGE THESE COLOURS -->
-    <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => deleteSelectedUsers()}>Yes</Button>
-    <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    <svelte:fragment slot="footer">
+        <Button class="bg-blue-200 hover:bg-blue-300 text-black" on:click={() => deleteSelectedUsers()}>Yes</Button>
+        <Button class="bg-red-200 hover:bg-red-300 text-black">No</Button>
+    </svelte:fragment>
 </Modal>
 
 <!-- figure out if this should be main -->
