+++ src/lib/InsertDataPage.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
 
-    import { Label, Input} from 'flowbite-svelte';
+    import { Label, Input, ButtonGroup} from 'flowbite-svelte';
     import { Dropdown, Search, Button } from 'flowbite-svelte';
     import { ChevronDownOutline } from 'flowbite-svelte-icons';
     import { Table, TableBody, TableBodyCell, TableBodyRow, TableHead, TableHeadCell } from 'flowbite-svelte';
@@ -10,6 +10,8 @@
     import AdminPage from './AdminPage.svelte';
 	import { invalidateAll } from '$app/navigation';
 	import ErrorMessage from './ErrorMessage.svelte';
+	import { error } from '@sveltejs/kit';
+	import { isNumber } from './utility';
 
     interface Props {
         employees?: Array<Employee>,
@@ -20,7 +22,7 @@
     }
 
     let { employees = [], showYears = true, showEmployeesDropdown = false, year, employee }: Props = $props();
-    
+
     const undefinedEmployee: Employee = {
         employeeID: "-1",
         firstName: "Undefined",
@@ -29,6 +31,11 @@
         dob: "Undefined",
         activeStatus: "Undefined"
     };
+
+    const undefinedEmployeeSearchable: EmployeeSearchable = $state({
+        name: "Select an employee",
+        data: undefinedEmployee
+    });
 
     // used to make it easier to access employees from their full name
     type EmployeeSearchable = {
@@ -44,10 +51,7 @@
         data: employee
     })) as Array<EmployeeSearchable>);
 
-    let selectedEmployee: EmployeeSearchable = $state({
-        name: "Select an employee", 
-        data: undefinedEmployee
-    });
+    let selectedEmployee: EmployeeSearchable = $state(undefinedEmployeeSearchable);
 
     let inputValueName: string = $state("");
 
@@ -60,66 +64,111 @@
         nameMenuOpen = false; 
     };
 
+    let showDataFields = $state((employee && year) ? true : false);
+
     let nameMenuOpen = $state(false);
 
     let inputValueYear = $state("");
     let leftFrequencies = $state({
-        hz500: '',
-        hz1000: '',
-        hz2000: '',
-        hz3000: '',
-        hz4000: '',
-        hz6000: '',
-        hz8000: ''
+        hz500: "",
+        hz1000: "",
+        hz2000: "",
+        hz3000: "",
+        hz4000: "",
+        hz6000: "",
+        hz8000: ""
     });
     let rightFrequencies = $state({
-        hz500: '',
-        hz1000: '',
-        hz2000: '',
-        hz3000: '',
-        hz4000: '',
-        hz6000: '',
-        hz8000: ''
+        hz500: "",
+        hz1000: "",
+        hz2000: "",
+        hz3000: "",
+        hz4000: "",
+        hz6000: "",
+        hz8000: ""
     });
 
     let success = $state(true);
-    let errorMessage = $state('');
-    let successMessage = $state('');
+    let errorMessage = $state("");
+    let successMessage = $state("");
 
     function displayError(message: string) {
         errorMessage = message;
         success = false;
     }
 
-    async function handleSubmit(event: Event) {
-        event.preventDefault(); // Prevent the default form submission behavior
-        successMessage = '';  
-        errorMessage = '';  
+    function preprocessFrequencies(frequencies: Record<string, string>) {
+        return Object.fromEntries(
+            Object.entries(frequencies).map(([key, value]) => [
+                key, 
+                value.trim().toUpperCase() === "CNT" ? null : value
+            ])
+        );
+    }
 
-        await addHearingData();
-
-        if (success) {
-            successMessage = 'Successfully added employee hearing data! Refreshing page...';
-            await invalidateAll(); // Refresh the page or data
-        }
-        else {
-            console.error('Error occurred: ', errorMessage);
+    async function checkYearAvailabilityKeydown(e: KeyboardEvent) {
+        if (e.key == "Enter") {
+            await checkYearAvailability();
         }
     }
 
-    function preprocessFrequencies(frequencies: Record<string, string>) {
-      return Object.fromEntries(
-          Object.entries(frequencies).map(([key, value]) => [
-              key, 
-              value.trim().toUpperCase() === "CNT" ? null : value
-          ])
-      );
+    const earliestYear = 1957;
+
+    async function checkYearAvailability() {
+        // Some checks to see if our ducks are in a row
+        if (!employee && selectedEmployee == undefinedEmployeeSearchable) {
+            displayError("No employee was selected!");
+            return;
+        }
+        if (!isNumber(inputValueYear)) {
+            displayError("The year is not a number...");
+            return;
+        }
+        const yearAsInteger = parseInt(inputValueYear);
+        if (yearAsInteger < earliestYear) {
+            displayError(`The selected year is before ${earliestYear}. Please choose a valid year.`);
+            return;
+        }
+        const currentYear = new Date().getFullYear();
+        if (yearAsInteger > currentYear) {
+            displayError(`The selected year is after ${currentYear}. Please choose a valid year.`);
+            return;
+        }
+        const formData = new FormData();
+
+        const appendedEmployeeID = employee ? employee.employeeID : selectedEmployee.data.employeeID;
+        formData.append('id', appendedEmployeeID); // Pass id
+        const appendedYear = year ? year : inputValueYear;
+        formData.append('year', appendedYear); // Year of data
+
+        const response = await fetch('/dashboard?/checkYearAvailability', {
+            method: 'POST',
+            body: formData,
+        });
+        showDataFields = false;
+        try {
+            console.log('Raw server response:', response);
+            const serverResponse = await response.json()
+            const result = JSON.parse(JSON.parse(serverResponse.data)[0]);
+
+            if (result["success"]) {
+                success = true;
+                showDataFields = true;
+            } else {
+                displayError(result["message"] ?? "Failed to add check available years.");
+                
+            }
+        } 
+        catch (error: any) {
+            console.error('Error during fetch or JSON parsing: ', error.message ?? 'no defined error message');
+            displayError(error.message ?? 'An error occurred');
+        }
     }
 
     async function addHearingData() {
         const formData = new FormData();
-        formData.append('leftEarFrequencies', JSON.stringify(leftFrequencies)); // Left ear data
-        formData.append('rightEarFrequencies', JSON.stringify(rightFrequencies)); // Right ear data
+        formData.append('leftEarFrequencies', JSON.stringify(preprocessFrequencies(leftFrequencies))); // Left ear data
+        formData.append('rightEarFrequencies', JSON.stringify(preprocessFrequencies(rightFrequencies))); // Right ear data
 
         const appendedEmployeeID = employee ? employee.employeeID : selectedEmployee.data.employeeID;
         formData.append('id', appendedEmployeeID); // Pass id
@@ -137,20 +186,18 @@
         try {
             // Debug: Log raw response
             console.log('Raw server response:', response);
-
             const serverResponse = await response.json()
-
             const result = JSON.parse(JSON.parse(serverResponse.data)[0]);
 
             if (result["success"]) {
                 success = true;
                 await invalidateAll();
             } else {
-                displayError(serverResponse.message ?? "Failed to add employee data.");
+                displayError(result["message"] ?? "Failed to add employee data.");
             }
         } 
         catch (error: any) {
-            console.error('Error during fetch or JSON parsing:', error);
+            console.error('Error during fetch or JSON parsing: ', error.message ?? 'no defined error message');
             displayError(error.message ?? 'An error occurred');
         }
     }
@@ -159,10 +206,7 @@
 
 <p class="center text-2xl">Add New Data</p>
 
-{#if !success}
-    <p class="mt-1 text-sm font-normal text-red-600 dark:text-red-300">{errorMessage}</p>
-{/if}
-
+<ErrorMessage {success} {errorMessage} />
 
 <div class="dropdown-container flex-container form"> 
     {#if showEmployeesDropdown}
@@ -189,64 +233,62 @@
         <!-- Add Year Input -->
         <div style="width: 300px; margin-left: 16px;">
             <Label for="year" class="block mb-2">Add Year</Label>
-            <Input id="year" placeholder="year" bind:value={inputValueYear} />
+            <ButtonGroup class="w-full">
+                <Input id="year" placeholder="1957" 
+                    bind:value={inputValueYear} on:keydown={checkYearAvailabilityKeydown} />
+                <Button color="blue" on:click={checkYearAvailability}>
+                    Check
+                </Button>
+            </ButtonGroup>
         </div>
     {/if}
 </div>
 
-    
-<Table> <!--  style="width: 90%; text-align: center; margin: auto;" -->
-    <TableHead>
-        <TableHeadCell></TableHeadCell>
-        <TableHeadCell>500 Hz</TableHeadCell>
-        <TableHeadCell>1000 Hz</TableHeadCell>
-        <TableHeadCell>2000 Hz</TableHeadCell>
-        <TableHeadCell>3000 Hz</TableHeadCell>
-        <TableHeadCell>4000 Hz</TableHeadCell>
-        <TableHeadCell>6000 Hz</TableHeadCell>
-        <TableHeadCell>8000 Hz</TableHeadCell>
-    </TableHead>
-    <TableBody tableBodyClass="divide-y">
-    <TableBodyRow>
-        <TableBodyCell>Left Ear</TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz500} placeholder="500" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz1000} placeholder="1000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz2000} placeholder="2000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz3000} placeholder="3000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz4000} placeholder="4000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz6000} placeholder="6000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={leftFrequencies.hz8000} placeholder="8000" required /></TableBodyCell>
-    </TableBodyRow>
-    <TableBodyRow>
-        <TableBodyCell>Right Ear</TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz500} placeholder="500" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz1000} placeholder="1000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz2000} placeholder="2000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz3000} placeholder="3000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz4000} placeholder="4000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz6000} placeholder="6000" required /></TableBodyCell>
-        <TableBodyCell><Input bind:value={rightFrequencies.hz8000} placeholder="8000" required /></TableBodyCell>
-    </TableBodyRow>
-    </TableBody>
-</Table>
+{#if showDataFields}
+    <Table> <!--  style="width: 90%; text-align: center; margin: auto;" -->
+        <TableHead>
+            <TableHeadCell></TableHeadCell>
+            <TableHeadCell>500 Hz</TableHeadCell>
+            <TableHeadCell>1000 Hz</TableHeadCell>
+            <TableHeadCell>2000 Hz</TableHeadCell>
+            <TableHeadCell>3000 Hz</TableHeadCell>
+            <TableHeadCell>4000 Hz</TableHeadCell>
+            <TableHeadCell>6000 Hz</TableHeadCell>
+            <TableHeadCell>8000 Hz</TableHeadCell>
+        </TableHead>
+        <TableBody tableBodyClass="divide-y">
+        <TableBodyRow>
+            <TableBodyCell>Left Ear</TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz500} placeholder="500" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz1000} placeholder="1000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz2000} placeholder="2000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz3000} placeholder="3000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz4000} placeholder="4000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz6000} placeholder="6000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={leftFrequencies.hz8000} placeholder="8000" required /></TableBodyCell>
+        </TableBodyRow>
+        <TableBodyRow>
+            <TableBodyCell>Right Ear</TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz500} placeholder="500" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz1000} placeholder="1000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz2000} placeholder="2000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz3000} placeholder="3000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz4000} placeholder="4000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz6000} placeholder="6000" required /></TableBodyCell>
+            <TableBodyCell><Input bind:value={rightFrequencies.hz8000} placeholder="8000" required /></TableBodyCell>
+        </TableBodyRow>
+        </TableBody>
+    </Table>
 
-<div class="form">
-    <Button 
-    class="bg-light-bluegreen hover:bg-dark-bluegreen text-black" 
-    style="width:200px" 
-    on:click={handleSubmit}
-    >Submit</Button>
-</div>
-
-<div>
-    {#if successMessage}
-        <div class="text-green-600 mt-4">
-        {successMessage}
-        </div>
-    {/if}
-
-    <ErrorMessage {success} {errorMessage} />
-</div>
+    <div class="form">
+        <Button 
+            class="bg-light-bluegreen hover:bg-dark-bluegreen text-black" 
+            style="width:200px" 
+            on:click={addHearingData}>
+            Submit
+        </Button>
+    </div>
+{/if}
 
 <style>
     /* TODO: Turn these into tailwind classes */
